Add files to OSF Storage if they are first dragged into the root folder and then dragged into the project folder
What was happening is that, once the file was seen by the desktop sync code, moving the file was considered a move, but moving only happens between projects. So I caught the case where it didn't really know where the file was being moved from and just said, well, if you don't know, then it must be a create action instead. I broke up the create into its own method that is separate from the event handler rather than trying to trigger a new event with the existing file.