Broadly ensure that errors on our HTTP requests will not kill our queue
This is, perhaps, a little excessive, but in keeping with the rest of the PR. It will reduce user feedback but ensure that things do not stay broken. If we have time over the next couple of days, handling this better will be a good idea, but it should at least allow QA a chance to test without killing the queue.