If node inaccessible, allow next to sync.
Fix issue that occurs when a user loses access to a project (because of event on the OSF- deletion, permissions revoked, etc).

An exception thrown at start of auditing (for project A) would prevent attempts to sync any other nodes after that in the list. Sync of project A would now be skipped, albeit silently and with no communication to the user.

This is a provisional solution. Improvements to user feedback and exception handling needed.