Handle Watchdog platform-dependent folder deletion.
Watchdog can't tell the difference between file and folder deletions on Windows- fortunately, our implementation for the two cases is the same, so for clarity the "fix" for this difference is to simply use the consistent implementation. [#OFF-62]

Also clarify an ambiguous source code comment and add logging statement if unexpected code is reached.