Explicitly flush the OperationWorker's queue when shutting down RemoteSyncworker
It was possible for the OperationWorker to be stopped and for the
RemoteSyncWorker to still try to push things into it's queue. This
caused the app to hang when trying to shut down mid-sync. Reordering
the shutdown sequence and explicitly waiting for the OperationWorker's
queue to be cleared during the RemoteSyncWorker's#stop fixes this.