Let LocalSyncWorker's init raise an exception when get_current_user returns None
This only happens when a user logs out and does not disrupt the user
experience.

TODO: Find a way to stop the Remote/Operation Workers without waiting
for their queues to be empty. When logging out the database gets dropped
and it would be better to just short circuit any in-progress sync
operations.